fix: - Chỉnh sửa cấu trúc layout giao diện chụp hình. - có thể lướt giữa màn hình chụp và màn hình thẻ icon, comment mượt mà. feat: - upload ảnh vừa chụp lên cloud. - Lưu ảnh vừa chụp vào firebase.
