+++ Locket/src/main/java/vn/edu/hcumuaf/locket/controller/UploadController.java
@@ -6,27 +6,25 @@ import io.netty.util.internal.ObjectUtil;
 import org.springframework.beans.MutablePropertyValues;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
 import java.util.Map;
 
 @RestController
-@RequestMapping("api/upload")
+@RequestMapping("/api/upload")
 public class UploadController {
 
     @Autowired
     private Cloudinary cloudinary;
 
-    @PostMapping
+    @CrossOrigin(origins = "*")
+    @PostMapping("/image")
     public ResponseEntity<?> uploadImage(@RequestParam("file") MultipartFile file) throws IOException {
         try {
             Map uploadResult = cloudinary.uploader().upload(file.getBytes(), ObjectUtils.emptyMap());
-            return ResponseEntity.ok(uploadResult.get("secure_url"));
+            return ResponseEntity.ok(Map.of("url", (String) uploadResult.get("secure_url")));
         } catch (IOException e) {
             return ResponseEntity.status(500).body("Lỗi upload: " + e.getMessage());
         }
