+++ Locket/src/main/resources/application.properties
@@ -11,7 +11,9 @@ logging.level.com.google.firebase=INFO
 # Async Configuration
 spring.task.execution.pool.core-size=10
 spring.task.execution.pool.max-size=20
-spring.task.execution.pool.queue-capacity=100 
+spring.task.execution.pool.queue-capacity=100
+spring.servlet.multipart.max-file-size=10MB
+spring.servlet.multipart.max-request-size=10MB
 
 spring.autoconfigure.exclude=org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration,org.springframework.boot.autoconfigure.orm.jpa.HibernateJpaAutoConfiguration
 spring.application.name=Locket
