+++ Locket_Mobile/app/src/main/AndroidManifest.xml
@@ -13,13 +13,13 @@
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
     <uses-permission android:name="android.permission.ACCESS_WIFI_STATE" />
     <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE"/>
-    <uses-permission android:name="android.permission.INTERNET" />
 
     <application
         android:name=".MyApplication"
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:fullBackupContent="@xml/backup_rules"
+        tools:replace="android:fullBackupContent"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
@@ -61,6 +61,7 @@
         <activity android:name=".activity.ShowImageActivity"/>
         <activity android:name=".activity.ListFriendActivity" />
         <activity android:name=".activity.HomeActivity" />
+        <activity android:name=".activity.PageComponentActivity" />
 <!--        <activity android:name=".MessageActivity"/>-->
 
     </application>
