+++ Locket_Mobile/app/src/main/java/com/hucmuaf/locket_mobile/modedb/SaveImageResponse.java
@@ -1,0 +1,8 @@
+package com.hucmuaf.locket_mobile.modedb;
+
+public class SaveImageResponse {
+    private String message;
+    public String getMessage(){
+        return message;
+    }
+}
