+++ Locket_Mobile/app/src/main/java/com/hucmuaf/locket_mobile/modedb/UploadImageResponse.java
@@ -1,0 +1,8 @@
+package com.hucmuaf.locket_mobile.modedb;
+
+public class UploadImageResponse {
+    private String url;
+    public String getUrl(){
+        return url;
+    }
+}
