+++ Locket_Mobile/app/src/main/java/com/hucmuaf/locket_mobile/service/ApiClient.java
@@ -10,7 +10,7 @@ import retrofit2.converter.gson.GsonConverterFactory;
 public class ApiClient {
     //     private static final String BASE_URL = "http://10.50.0.1:8080/"; // For Android emulator
 
-    private static final String BASE_URL = "http://192.168.1.166:8080/"; // For real device, your computer's IP
+    private static final String BASE_URL = "http://192.168.78.24:8080/"; // For real device, your computer's IP
     // private static final String BASE_URL = "http://localhost:8080/"; // For testing
