+++ Locket_Mobile/app/src/main/java/com/hucmuaf/locket_mobile/service/ImageService.java
@@ -1,11 +1,16 @@
 package com.hucmuaf.locket_mobile.service;
 
 import com.hucmuaf.locket_mobile.modedb.Image;
+import com.hucmuaf.locket_mobile.modedb.SaveImageResponse;
 
 import java.util.List;
+import java.util.Map;
 
+import okhttp3.ResponseBody;
 import retrofit2.Call;
+import retrofit2.http.Body;
 import retrofit2.http.GET;
+import retrofit2.http.POST;
 import retrofit2.http.Path;
 
 public interface ImageService {
@@ -17,5 +22,8 @@ public interface ImageService {
 
     @GET("api/images/user/{userId}")
     Call<List<Image>> getImageOfUser(@Path("userId") String userId);
+
+    @POST("api/images/save")
+    Call<SaveImageResponse> saveImage(@Body Image image);
 
 }
