+++ Locket_Mobile/app/src/main/java/com/hucmuaf/locket_mobile/service/UploadImageService.java
@@ -1,0 +1,16 @@
+package com.hucmuaf.locket_mobile.service;
+
+import com.hucmuaf.locket_mobile.modedb.UploadImageResponse;
+
+import okhttp3.MultipartBody;
+import okhttp3.ResponseBody;
+import retrofit2.Call;
+import retrofit2.http.Multipart;
+import retrofit2.http.POST;
+import retrofit2.http.Part;
+
+public interface UploadImageService {
+    @Multipart
+    @POST("/api/upload/image")
+    Call<UploadImageResponse> uploadImage(@Part MultipartBody.Part file);
+}
