+++ Locket_Mobile/app/src/main/res/layout/home.xml
@@ -17,16 +17,9 @@
         android:layout_height="60dp"
         android:layout_marginTop="30dp"
         android:orientation="horizontal"
+        android:gravity="center_horizontal"
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintStart_toStartOf="parent">
-
-        <ImageView
-            android:id="@+id/account"
-            android:layout_width="50dp"
-            android:layout_height="50dp"
-            android:layout_marginStart="17dp"
-            android:background="@drawable/circle_border"
-            android:src="@mipmap/account_icon" />
 
         <LinearLayout
             android:id="@+id/friends"
@@ -47,25 +40,29 @@
                 android:src="@mipmap/groups"
                 android:clickable="false"
                 android:focusable="false" />
+            <TextView
+                android:id="@+id/count_friend"
+                android:layout_width="wrap_content"
+                android:layout_height="30dp"
+                android:gravity="center_horizontal"
+                android:layout_marginStart="10dp"
+                android:text="0"
+                android:textColor="#ffffff"
+                android:textSize="20dp"
+                android:clickable="false"
+                android:focusable="false" />
 
             <TextView
-                android:layout_width="100dp"
+                android:layout_width="80dp"
                 android:layout_height="30dp"
-                android:layout_marginStart="5dp"
                 android:gravity="center"
-                android:text="6 Bạn bè"
+                android:text="Bạn bè"
                 android:textColor="#ffffff"
                 android:textSize="20dp"
                 android:clickable="false"
                 android:focusable="false" />
         </LinearLayout>
 
-        <ImageView
-            android:id="@+id/message"
-            android:layout_width="50dp"
-            android:layout_height="50dp"
-            android:background="@drawable/circle_border"
-            android:src="@mipmap/message_circle" />
     </LinearLayout>
 
     <FrameLayout
