+++ Locket_Mobile/app/src/main/res/layout/react_emoji_comment.xml
@@ -31,15 +31,8 @@
             android:layout_height="60dp"
             android:layout_marginTop="30dp"
             android:elevation="1dp"
-            android:orientation="horizontal">
-
-            <ImageView
-                android:id="@+id/account"
-                android:layout_width="50dp"
-                android:layout_height="50dp"
-                android:layout_marginStart="17dp"
-                android:background="@drawable/circle_border"
-                android:src="@mipmap/account_icon" />
+            android:orientation="horizontal"
+            android:gravity="center_horizontal">
 
             <LinearLayout
                 android:id="@+id/all_friends"
@@ -67,12 +60,6 @@
                     android:src="@mipmap/down_toggle" />
             </LinearLayout>
 
-            <ImageView
-                android:id="@+id/message"
-                android:layout_width="50dp"
-                android:layout_height="50dp"
-                android:background="@drawable/circle_border"
-                android:src="@mipmap/message_circle" />
         </LinearLayout>
