+++ Locket_Mobile/app/src/main/res/xml/network_security_config.xml
@@ -15,8 +15,8 @@
         <domain includeSubdomains="true">192.168.1.109</domain>
         <domain includeSubdomains="true">192.168.1.110</domain>
         <domain includeSubdomains="true">10.0.185.202</domain>
-        <domain includeSubdomains="true">10.0.208.94</domain>
-        <domain includeSubdomains="true">192.168.1.166</domain>
+        <domain includeSubdomains="true">10.51.33.182</domain>
+        <domain includeSubdomains="true">192.168.78.24</domain>
     </domain-config>
 </network-security-config>
